(bluefox) fix error with loading of monaco-editor
Because system tries to load http://localhost:8081/adapter/javascript/monaco-editor//editor/editor.main.js (//) and this does not work if files stored in redis.
Another error is that typescript worker was loaded always from vs/language/typescript/tsWorker.js and not from monaco-editor/language/typescript/tsWorker.js. May be this is bug in monaco loader. But it could be fixed by renaming of the folder.