fix cacheObjectEnums
Cache of enum objects was broken. The "return" for parent prevented it to be filled for every child objects at all. If for all parts in a path objects were defined (i.e. instance, device, channel, ..., state), then the cache only got entries for the instance.

By handing the child enumId arrays to the parent, those cache entries were filled with the enums of the first child object processed (which is random). That explains the errounous behaviour that some users (including me) experienced.

Also the cache was declared locally and therefore never cleared. I.e. changes in enums would not be reflected. There was a global cache that already is cleared in onObjectChange in main.js, but it was never used.

Fixes #523 (tested by me) and might fix #515.
Might also fixes #466 , #364